feat: move task logic from mod.rs to task.rs (#569)
