+++ dragonfly-client/src/bin/dfdaemon/main.rs
@@ -25,9 +25,9 @@ use dragonfly_client::grpc::{
 use dragonfly_client::health::Health;
 use dragonfly_client::metrics::Metrics;
 use dragonfly_client::proxy::Proxy;
+use dragonfly_client::resource::task::Task;
 use dragonfly_client::shutdown;
 use dragonfly_client::stats::Stats;
-use dragonfly_client::task::Task;
 use dragonfly_client::tracing::init_tracing;
 use dragonfly_client_backend::BackendFactory;
 use dragonfly_client_config::dfdaemon;
