+++ dragonfly-client/src/resource/piece_collector.rs
